requirements.txt
numpy==1.19.2
torch==1.5.0
torchvision==0.6.0
scipy==1.4.1
scikit-image==0.17.2
tqdm==4.50.2
opencv-python==4.4.0.46
matplotlib==3.3.2
Pillow==8.0.1
